==> notebooks/1_water_potability_exploration.ipynb <==
# Water Potability Exploratory Notebook

import pyspark
import pyspark.sql.functions as F
import pyspark.sql.types as T
import sys
from pyspark.sql import SparkSession, Window, DataFrame
from pyspark.mllib.stat import Statistics

sys.path.append('/home/jovyan/work')

spark = SparkSession.builder.getOrCreate()

## Step 1 - Exploration

df = spark.read.csv('../data/water_potability.csv',inferSchema=True, header=True)

print('Total record count: {}'.format(df.count()))

### Make a Train/Test Split

df_train, df_test = df.randomSplit([0.7, 0.3], seed=42)

# Save the train and test datasets
df_train.toPandas().to_csv('../data/water_potability_train.csv')

df_test.toPandas().to_csv('../data/water_potability_test.csv')

# Get rid of df so we don't accidentally use it
del df

df_train.printSchema()

print('Train record count: {}'.format(df_train.count()))

### View Sample Rows

df_train.show(5)

So, this is a binary classification problem with target values of notckd (not chronic kidney disease) and ckd (chronic kidney disease). **Note:** There is an extraneous tab (\t) on at lease one of the target values. We will have to fix that up.

### Target Column Distribution

freq_table = df_train.select(F.col("Potability").cast("integer")).groupBy("Potability").count().toPandas()
freq_table

This is a binary classification problem with a somewhat imbalanced dataset.

### Summary Statistics

df_train.describe().toPandas().transpose()

### Null Counts

for col in df_train.columns:
    print('{} Null Count: {}'.format(col, df_train.where(F.col(col).isNull()).count()))

We have nulls to deal with for pH, sulfates, and trihalomethanes.

## Step 2 - Pipeline Development

from pyspark.ml.feature import Imputer
from pyspark.ml import Pipeline
from pyspark.ml.feature import VectorAssembler, PolynomialExpansion, StandardScaler

### Impute Missing Values

imputer = Imputer(
    inputCols=['ph', 'Sulfate', 'Trihalomethanes'],
    outputCols=['ph_imp', 'Sulfate_imp', 'Trihalomethanes_imp']
)
df_imputed = imputer.setStrategy("mean").fit(df_train).transform(df_train)
df_imputed = df_imputed.select('Hardness','Solids','Chloramines','Conductivity',
                              'Organic_carbon','Turbidity','ph_imp', 'Sulfate_imp', 
                              'Trihalomethanes_imp', 'Potability')
df_imputed.show(10)

### Create Feature Vector

vec_assembler = VectorAssembler(
    inputCols=['Hardness','Solids','Chloramines','Conductivity','Organic_carbon',
               'Turbidity','ph_imp', 'Sulfate_imp', 'Trihalomethanes_imp'],
    outputCol='Features'
)
df_features = vec_assembler.transform(df_imputed)
df_features.select('Features').take(1)

### Scale the Features

scaler = StandardScaler(
    inputCol='Features', outputCol='ScaledFeatures',
    withStd=True, withMean=True
)

scalerFit = scaler.fit(df_features)

df_features_scaled = scalerFit.transform(df_features)
df_features_scaled.select('ScaledFeatures').take(1)

### Perform Polynomial Feature Expansion

poly_feature_exp = PolynomialExpansion(degree=3, inputCol="ScaledFeatures", outputCol="PolynomialFeatures")
poly_features = poly_feature_exp.transform(df_features_scaled)
poly_features.select('PolynomialFeatures').take(1)
