==> notebooks/2_water_potability_pipeline_development.ipynb <==
# Water Potability Pipeline Development Notebook

import sys
sys.path.append('/home/jovyan/work')

from handyspark import *

import pyspark
import pyspark.sql.functions as F
import pyspark.sql.types as T
from pyspark.sql import SparkSession, DataFrame
from pyspark.ml.feature import Imputer, VectorAssembler, PolynomialExpansion, StandardScaler

# from pyspark.mllib.stat import Statistics
#from pipeline import feature_pipeline

spark = SparkSession.builder.getOrCreate()
spark.sparkContext.setLogLevel('ERROR')

## Step 2 - Feature Pipeline Development

### Make Test-Train Split

It is important to make the test train split before doing any feature pipeline development. Giving your pipline prior knowledge of the test data is a form of data leakage that can introduce bias into the model.

# Randomly split the the dataset into test and train.
# Set the seed value for reproducability

df = spark.read.csv('../data/water_potability.csv',inferSchema=True, header=True)
df_train, df_test = df.randomSplit([0.8, 0.2], seed = 42)
print("Training Dataset Count: " + str(df_train.count()))
print("Test Dataset Count: " + str(df_test.count()))

### Impute Missing Values

Imputing missing values based on the value of the target column is somewhat difficult to do with Pyspark, so I am going to use a package called HandySpark which makes it easier.

hdf = df_train.toHandy()


#hdf_filled = hdf.stratify(['Potability']).fill(continuous=['ph'], strategy=['mean'])
#df_train = hdf_filled.notHandy()

# Check for Missing Data
#df_train.select(*(F.sum(F.col(c).isNull().cast("int")).alias(c) for c in df.columns)).show()

#imputer = Imputer(
#    inputCols=['ph', 'Sulfate', 'Trihalomethanes'],
#    outputCols=['ph', 'Sulfate', 'Trihalomethanes']
#)
#df_imputed = imputer.setStrategy("mean").fit(df_train).transform(df_train)
#df_imputed.show(10)

### Create Feature Vector

vec_assembler = VectorAssembler(
    inputCols=['Hardness','Solids','Chloramines','Conductivity','Organic_carbon',
               'Turbidity'],
    outputCol='Features'
)
df_features = vec_assembler.transform(df_imputed)
df_features.select('Features').take(3)

### Scale the Features

scaler = StandardScaler(
    inputCol='Features', outputCol='ScaledFeatures',
    withStd=True, withMean=True
)

scalerFit = scaler.fit(df_features)

df_features_scaled = scalerFit.transform(df_features)
df_features_scaled.select('ScaledFeatures').take(3)

### Perform Polynomial Feature Expansion

poly_feature_exp = PolynomialExpansion(degree=3, inputCol="ScaledFeatures", outputCol="PolynomialFeatures")
poly_features = poly_feature_exp.transform(df_features_scaled)
poly_features.select('PolynomialFeatures').take(1)

## Test Pipeline Code

feature_cols = ['ph','Hardness','Solids','Chloramines','Sulfate','Conductivity',
               'Organic_carbon', 'Trihalomethanes', 'Turbidity']
assembler_out_col = 'Features'
scaler_out_col = 'ScaledFeatures'
expander_out_col = 'ExpandedFeatures'
degree = 3

pipeline = feature_pipeline.create_feature_pipeline(df_train, feature_cols, assembler_out_col, 
                                                    scaler_out_col, expander_out_col, degree)

pipeline.transform(df_train).select('ExpandedFeatures').take(1)
