==> notebooks/3_water_potability_model_exploration.ipynb <==
# Water Potability Model Exploration

import sys
sys.path.append('/home/jovyan/work')

import mlflow
import mlflow.spark
import mlflow.mleap

import mleap.pyspark

from pyspark.sql import SparkSession,DataFrame
from pyspark.ml import PipelineModel
from pyspark.ml.evaluation import BinaryClassificationEvaluator
from pyspark.ml.classification import GBTClassifier

from pipeline import feature_pipeline_builder

import os

spark = SparkSession.builder  \
    .config('spark.jars.packages', 'ml.combust.mleap:mleap-spark-base_2.11:0.17.0,ml.combust.mleap:mleap-spark_2.11:0.17.0') \
    .appName('ClassifierTraining')  \
    .getOrCreate()

## Load the Training Data

file_path = '../data/water_potability_train.csv'
df_train = spark.read.csv(file_path,inferSchema=True, header=True)
df_train = df_train.drop('_c0')
df_train.show(5)

## Load the Test Data

file_path = '../data/water_potability_test.csv'
df_test = spark.read.csv(file_path,inferSchema=True, header=True)
df_test = df_test.drop('_c0')
df_test.show(5)

## Create the Feature Pipeline

degree = 3
prediction_col = 'Potability'
feature_cols = ['ph','Hardness','Solids','Chloramines','Sulfate','Conductivity',
               'Organic_carbon', 'Trihalomethanes', 'Turbidity']

assembler_out_col = 'feature_vector'
scaler_out_col = 'scaled_features'
expander_out_col = 'features'

feature_pipeline = feature_pipeline_builder.create_feature_pipeline(feature_cols,
                                                    assembler_out_col,
                                                    scaler_out_col,
                                                    expander_out_col, 
                                                    degree)

feature_model = feature_pipeline.fit(df_train)

### Transform the Training Set

train_with_features = feature_model.transform(df_train)
train_expanded_features = train_with_features.select(prediction_col,expander_out_col)
train_persisted = train_expanded_features.persist()
train_expanded_features.show(5)

## Model Experimentation with MLFlow
### Set up Experiment

experiment = 'water_potability'
mlflow.set_tracking_uri(os.environ['MLFLOW_TRACKING_URI'])
mlflow.set_experiment(experiment)

### Fit the Model to the Training Set

gbt = GBTClassifier(featuresCol=expander_out_col,labelCol=prediction_col, maxIter=10)
fit_model = gbt.fit(train_persisted)

### Transform the Test Set

test_with_features = feature_model.transform(df_test)
test_expanded_features = test_with_features.select(prediction_col, expander_out_col)
test_persisted = test_expanded_features.persist()
test_expanded_features.show(5)

### Evaluate Model Performance

results = fit_model.transform(test_persisted)

my_eval = BinaryClassificationEvaluator(rawPredictionCol='prediction',
                                       labelCol=prediction_col)
results.select(prediction_col,'prediction')
AUC = my_eval.evaluate(results)
print("AUC score is : ",AUC)

## Add GBTClassifier to Pipeline Model and Save

final_pipeline = PipelineModel(stages = [feature_model , gbt])

mlflow.spark.save_model(spark_model=final_pipeline,
                        path='../models/water_potability_model')
